requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
Django==4.0.2
idna==3.3
pyTelegramBotAPI==4.4.0
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
